Prey and predator simulation
Simulation of behaviour of 2 populations. One of prey which are named "fish" where they multiply and "Sharks" which eat fish and multiply only when they have food. THe goal was to show that with specific parameters, the whole ecosystem will reach equilibrium where the number of prey and predators stay stable.